Tweaks
Tweaked stance speed and sprint penalty
